The real-time clock has to survive being added up
Setting the rate to a day per day did not make the clock work. At 1/86400 days per
second the sim adds 1.9e-7 days a tick, and an f32 holding an age of four days cannot
represent a step that small: the add rounds home and the age stops moving. Measured —
ages ran 24% fast between one and four days, then froze dead at four. No worker ever
reached NURSE_UNTIL, so every ant stayed a nurse, nothing dug, and nobody died of old
age. The clock said real time and the colony was a still photograph.

Every accumulator of colony-days is f64 now: ant and brood ages, the lay clock, the
rate itself, the stage and lifespan constants. The save format is JSON, so the wider
ant age needs no version bump.

Two tests, and the first one starts the sum at thirty days rather than zero, which is
the whole point of it.

+++ src/devcapture.rs
@@ -565,11 +565,11 @@ const C_QUIT: f32 = 125.0;
 
 /// Colony-days per real second for the scripted runs.
 ///
-/// The game runs at a colony day per real hour, so a two-minute capture covers three
-/// hundredths of a day and nothing about the brood is observable at all. At one day a second
+/// The game runs in real time — a colony day per day — so a two-minute capture covers a
+/// thousandth of a day and nothing about the brood is observable at all. At one day a second
 /// the same run covers a hundred and twenty-five days — twenty brood cycles — which is the
 /// only way a two-minute test can say anything about a six-day life stage.
-const CAPTURE_DAYS_PER_SECOND: f32 = 1.0;
+const CAPTURE_DAYS_PER_SECOND: f64 = 1.0;
 
 pub fn run_colony_capture(
     mut commands: Commands,
